pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["src/caul"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv-dynamic-versioning]
fallback-version = "0.0.0"

[project]
name = "caul"
dynamic = ["version"]
description = "Python implementation of an ASR service"
authors = [
    {name = "Lion Summerbell", email = "lsummerbell@icij.org"}
]
requires-python = ">=3.11.0, <3.14.0"
dependencies = [
    "numpy~=2.3",
    "torchaudio~=2.10",
    "torchcodec>=0.10.0,<0.12",
    "icij-common~=0.8.1",
    "kaldialign~=0.9",
    "langcodes~=3.5",
    "pydantic-extra-types[pycountry]~=2.11",
]

[project.optional-dependencies]
nemo = [
  "nemo-toolkit[asr]~=2.7",
  "torch~=2.10",
]
pyannote = [
    "pyannote-audio>=4.0.4",
]
fireredasr2 = [
  "fireredasr2s @ git+https://github.com/FireRedTeam/FireRedASR2S.git",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" }
]
torchaudio = [
  { index = "pytorch-cpu" }
]
torchcodec = [
  { index = "pytorch-cpu" }
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pre-commit>=4.5.1",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
]

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]

[tool.pylint.main]
fail-under = 8.89

[tool.pylint.refactoring]
max-nested-blocks = 5
never-returning-functions = ["sys.exit", "argparse.parse_error"]

[tool.pylint.similarities]
# Docstrings are removed from the similarity computation
ignore-docstrings = true
# Comments are removed from the similarity computation
ignore-comments = true

[tool.black]
target-version = ["py310"]
